requirements-gcp.txt
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
bcrypt==4.0.1
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.3.2
colorama==0.4.6
constantly==23.10.4
cryptography==42.0.8
daphne==4.0.0
defusedxml==0.7.1
diff-match-patch==20230430
distro==1.9.0
Django==4.2.1
django-cleanup==8.0.0
django-environ==0.11.2
django-extra-views==0.14.0
django-import-export==3.3.7
django-redis==5.4.0
django-searchable-encrypted-fields==0.2.1
django-storages==1.14.2
et-xmlfile==1.1.0
exceptiongroup==1.2.1
google-api-core==2.19.0
google-auth==2.30.0
google-cloud-aiplatform==1.38.1
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-secret-manager==2.18.2
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.1
grpc-google-iam-v1==0.13.0
grpcio==1.64.1
grpcio-status==1.62.2
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
MarkupPy==1.14
msgpack==1.0.8
mysqlclient==2.2.4
numpy==1.26.4
oauthlib==3.2.2
odfpy==1.4.1
openai==1.3.8
openpyxl==3.1.4
packaging==24.1
pandas==2.1.2
pillow==10.3.0
proto-plus==1.24.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.4
pydantic_core==2.18.4
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.6
regex==2024.5.15
requests==2.32.3
requests-oauthlib==1.3.1
rsa==4.9
service-identity==24.1.0
shapely==2.0.4
six==1.16.0
sniffio==1.3.1
sorl-thumbnail==12.10.0
sqlparse==0.5.0
tablib==3.5.0
tiktoken==0.7.0
tqdm==4.66.4
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
Unidecode==1.3.7
urllib3==2.2.2
xlrd==2.0.1
xlwt==1.3.0
zope.interface==6.4.post2
